+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.3] - 2024-07-28
+
+### Changed
+
+- Updated dependencies to use paper-api instead of spigot-api for event needed to check when a player loads a crafting
+  recipe.
+
+### Fixed
+
+- Fixed shulker box limit ignored when loading crafting recipe.
+
+
 ## [1.2.2] - 2024-07-17
 
 ### Fixed